Updated routing & context
