+++ apps/console/app/routes/__layout/groups/create.tsx
@@ -1,0 +1,38 @@
+import { generateTraceContextHeaders } from '@proofzero/platform-middleware/trace'
+import { AccountURN } from '@proofzero/urns/account'
+import { getRollupReqFunctionErrorWrapper } from '@proofzero/utils/errors'
+import { ActionFunction, redirect } from '@remix-run/cloudflare'
+import { parseJwt, requireJWT } from '~/utilities/session.server'
+import createAccountClient from '@proofzero/platform-clients/account'
+import { getAuthzHeaderConditionallyFromToken } from '@proofzero/utils'
+import { BadRequestError } from '@proofzero/errors'
+
+export const action: ActionFunction = getRollupReqFunctionErrorWrapper(
+  async ({ request, context }) => {
+    const jwt = await requireJWT(request)
+    const parsedJwt = parseJwt(jwt!)
+    const accountURN = parsedJwt.sub as AccountURN
+
+    const traceHeader = generateTraceContextHeaders(context.traceSpan)
+
+    const accountClient = createAccountClient(Account, {
+      ...getAuthzHeaderConditionallyFromToken(jwt),
+      ...traceHeader,
+    })
+
+    const fd = await request.formData()
+    const name = fd.get('name')
+    if (!name) {
+      throw new BadRequestError({
+        message: 'Name is required',
+      })
+    }
+
+    await accountClient.createIdentityGroup.mutate({
+      accountURN,
+      name: name as string,
+    })
+
+    return redirect('/groups')
+  }
+)
